Improve patch release detection after publishing package fix versions
A forced package update with a 3-part version number will convert the
version to a 4-part number. When DigiDNA creates patch releases with
the same 3-part prefix, AU would not create a new package due to the
patch fix version being greater.

To work around this, storing a local cache of the latest consumed
software version for each stream (which will not necessarily equate to
the package version) when updating. When the latest package version
happens to be a package fix version, double-check against this cache to
determine whether we need to force an update.